requirements.txt
anyio==3.6.2
api==0.0.7
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
distlib==0.3.1
Django==3.1.4
filelock==3.0.12
Flask==1.1.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==2.10
importlib2==3.5.0.2
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
linecache2==1.0.0
MarkupSafe==1.1.1
more-itertools==8.6.0
nose==1.3.7
numpy==1.24.0
packaging==20.8
pandas==1.5.2
peewee==3.14.0
pluggy==1.0.0.dev0
py==1.10.0
pyclean==2.2.0
pyparsing==2.4.7
pyTelegramBotAPI==4.8.0
pytest==6.2.1
python-dateutil==2.8.2
python-telegram-bot==20.0b0
pytz==2020.4
requests==2.25.1
rfc3986==1.5.0
six==1.15.0
sniffio==1.3.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
traceback2==1.4.0
tzlocal==2.1
urllib3==1.26.2
virtualenv==20.2.2
wcwidth==0.2.5
Werkzeug==1.0.1
xmltodict==0.12.0
